pyproject.toml
[project]
name = "agent-army"
version = "0.1.0"
description = "Enterprise Sales Organization Platform with Multi-Agent AI Crew"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.18.0",
    "redis[hiredis]>=5.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # LLM (for next plan)
    "litellm>=1.60.0",
    # HTTP
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    # Logging
    "structlog>=24.0.0",
    # Monitoring (for plan 03)
    "prometheus-client>=0.21.0",
    "sentry-sdk[fastapi]>=2.0.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
